+++ crates/router/src/types.rs
@@ -1321,6 +1321,10 @@ impl ForeignFrom<&SetupMandateRouterData> for PaymentsAuthorizeData {
             mit_category: None,
             billing_descriptor: data.request.billing_descriptor.clone(),
             tokenization: None,
+            partner_merchant_identifier_details: data
+                .request
+                .partner_merchant_identifier_details
+                .clone(),
         }
     }
 }